tidying and fixing

+++ debug.c
@@ -256,13 +256,13 @@ void print_mips(MIPS *mips, FILE *file)
       fprintf(file, "%s label%d\n", get_instruction(mips->instruction), mips->operand_one);
       break;
     case JUMPTOFUNC:
-      fprintf(file, "%s %s\n", get_instruction(mips->instruction), ((TOKEN*)((LOCATION*)mips->operand_one)->token)->lexeme);
+      fprintf(file, "%s %s\n", get_instruction(mips->instruction), ((TOKEN*)((LOCATION*)(long)mips->operand_one)->token)->lexeme);
       break;
     case JUMPTOADDRS:
-      fprintf(file, "%s %s\n", get_instruction(mips->instruction), ((TOKEN*)((LOCATION*)mips->operand_one)->token)->lexeme);
+      fprintf(file, "%s %s\n", get_instruction(mips->instruction), ((TOKEN*)((LOCATION*)(long)mips->operand_one)->token)->lexeme);
       break;
     case FUNCTION_DEF:
-      fprintf(file, "%s:\n", ((TOKEN*)((LOCATION*)mips->operand_one)->token)->lexeme);
+      fprintf(file, "%s:\n", ((TOKEN*)((LOCATION*)(long)mips->operand_one)->token)->lexeme);
       break;
     case SYSCALL:
       fprintf(file, "syscall\n");
